Add files via upload
Fix a bug in REC_PACK state - a collided packet which is not destined t the device.
Run some experiments with 4 devices, standard and APD-based ST - good results in non-fixed sized packet.
There is still a bug in constant-sized packets - packets are sent multiple times with curRet=0, we have to fix it! And also the packetsDS does not behave properly.